Move parse command logic into parse submodule
